Move plating and sequencing files
Moved to experiment/, which will house files relating to experimental
design and execution. Note, the R files in plating-and-sequencing/ refer
to old paths and versions of the sample data and would need to edited or
re-run from an earlier commit if it is desired to regenerate the output
files in this directory.