Create phyloseq object for initial bias analysis
The code in "code/2020-09-12-import-to-phyloseq.R" creates a phyloseq
object with the current abundance data from A1, S1, and S2 for a
simplified feature set corresponding to just the main experimental
strains - the inoculum strains and the T. mobilis control. This object
is stored in output/phyloseq, which contains a Readme to explain its
use.

As a usage test, I updated the document 2020-08-19-report.Rmd to use
this new phyloseq object.