Add PBS and aliquot volumes to sample data
Different volumes of PBS used for the Batch-2 inoculum and T. mobilis
samples than in Batch 1, in an effort to increase the cell
concentration. See the 2019-12-11 Wednesday notebook entries for
details. I'm adding these volumes to the specimen data table so that we
can account for this variation when analyzing extraction yields.

I added the volumes to the source Google spreadsheet, and edited and
re-ran the R script data/sample-data/00-fetch-from-drive.R to update the
CSV file containing the raw specimen metadata. An entry with the aliquot
volumes is also added; however these volumes are 200 uL for all
specimens.

I then edited and re-ran code/2020-08-14-prepare-sample-data.R, which
cleans and saves the DNA and sequenced sample metadata.